+++ SUI/SUI/ClimateSliderViewModel.swift
@@ -1,0 +1,9 @@
+// ClimateSliderViewModel.swift
+// Copyright © RoadMap. All rights reserved.
+
+import Foundation
+
+/// Вью модель ActionSheet
+final class ClimateSliderViewModel: ObservableObject {
+    // MARK: - Public properties
+}
