+++ SUI/SUI/NeumorphismSelected.swift
@@ -5,6 +5,8 @@ import SwiftUI
 
 /// Стиль нейморфизм выбранный
 struct NeumorphismSelected: ViewModifier {
+    // MARK: - Public property
+
     func body(content: Content) -> some View {
         content
             .shadow(color: .lightShadow, radius: 5, x: 5, y: 5)
