+++ SUI/SUI/NeumorphismUnSelected.swift
@@ -5,6 +5,8 @@ import SwiftUI
 
 /// Стиль нейморфизм не выбранный
 struct NeumorphismUnSelected: ViewModifier {
+    // MARK: - Public property
+
     func body(content: Content) -> some View {
         content
             .shadow(color: .lightShadow, radius: 5, x: -5, y: -5)
