+++ SUI/SUI/NeumorphismUnSelectedCircle.swift
@@ -5,12 +5,20 @@ import SwiftUI
 
 /// Стиль нейморфизм не выбранный круг
 struct NeumorphismUnSelectedCircle: ViewModifier {
+    // MARK: - Private Constants
+
+    private enum Constants {
+        static let backgroundColorString = "backgroundLight"
+    }
+
+    // MARK: - Public property
+
     func body(content: Content) -> some View {
         content
             .padding(.all, 10)
             .background(
                 Circle()
-                    .fill(Color("backgroundLight"))
+                    .fill(Color(Constants.backgroundColorString))
             )
             .neumorphismUnSelectedStyle()
     }
